# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API que intermedia a comunicação entre WhatsApp (via NeoHub) e WatsonX Assistant.",
        "title": "WhatsApp WatsonX API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/watsonx/message": {
            "post": {
                "description": "Sends a text message to WatsonX Assistant and returns the response (bypasses WhatsApp flow)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watsonx"
                ],
                "summary": "Send message directly to WatsonX",
                "parameters": [
                    {
                        "description": "Message payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DirectWatsonXRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.WatsonXDirectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/watsonx/session": {
            "post": {
                "description": "Creates a new session with Watson Assistant and returns the session ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watsonx"
                ],
                "summary": "Create a new Watson session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreateSessionResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/watsonx/sessions/reset": {
            "delete": {
                "description": "Deletes all cached Watson sessions from Redis",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watsonx"
                ],
                "summary": "Reset all sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/watsonx/sessions/stats": {
            "get": {
                "description": "Returns information about cached sessions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watsonx"
                ],
                "summary": "Get session cache statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/webhook/meta": {
            "post": {
                "description": "Receives messages from Meta WhatsApp Business API, processes through Watson, and sends reply back",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhook"
                ],
                "summary": "Webhook for Meta WhatsApp Business API",
                "parameters": [
                    {
                        "description": "Meta webhook payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MetaWebhookPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        },
        "/whatsapp/send": {
            "post": {
                "description": "Sends a text message to a WhatsApp number via NeoHub broker (bypasses WatsonX flow)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "whatsapp"
                ],
                "summary": "Send message directly to WhatsApp",
                "parameters": [
                    {
                        "description": "Message payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DirectWhatsAppRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "message sent"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "models.CreateSessionResponse": {
            "type": "object",
            "properties": {
                "session_id": {
                    "type": "string",
                    "example": "abc123"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "models.DirectWatsonXRequest": {
            "type": "object",
            "properties": {
                "session_id": {
                    "type": "string",
                    "example": ""
                },
                "text": {
                    "type": "string",
                    "example": "Olá, preciso de ajuda"
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "models.DirectWhatsAppRequest": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string",
                    "example": "Olá, esta é uma mensagem de teste"
                },
                "to": {
                    "type": "string",
                    "example": "5511999999999"
                }
            }
        },
        "models.MetaChange": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "value": {
                    "$ref": "#/definitions/models.MetaValue"
                }
            }
        },
        "models.MetaContact": {
            "type": "object",
            "properties": {
                "profile": {
                    "$ref": "#/definitions/models.MetaProfile"
                },
                "wa_id": {
                    "type": "string"
                }
            }
        },
        "models.MetaEntry": {
            "type": "object",
            "properties": {
                "changes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MetaChange"
                    }
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "models.MetaMessage": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "text": {
                    "$ref": "#/definitions/models.MetaTextBody"
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.MetaMetadata": {
            "type": "object",
            "properties": {
                "display_phone_number": {
                    "type": "string"
                },
                "phone_number_id": {
                    "type": "string"
                }
            }
        },
        "models.MetaProfile": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.MetaTextBody": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        },
        "models.MetaValue": {
            "type": "object",
            "properties": {
                "contacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MetaContact"
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MetaMessage"
                    }
                },
                "messaging_product": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.MetaMetadata"
                }
            }
        },
        "models.MetaWebhookPayload": {
            "type": "object",
            "properties": {
                "entry": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MetaEntry"
                    }
                },
                "object": {
                    "type": "string"
                }
            }
        },
        "models.WatsonXDirectResponse": {
            "type": "object",
            "properties": {
                "reply": {
                    "type": "string",
                    "example": "Olá! Como posso ajudar?"
                },
                "session_id": {
                    "type": "string",
                    "example": "abc123"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        }
    }
}